修改概率的位数

+++ historyLotto/src/main/java/com/lsy/historylotto/entity/SummaryStatisticsItemEntity.kt
@@ -8,5 +8,5 @@ package com.lsy.historylotto.entity
 data class SummaryStatisticsItemEntity(
     var title: String,
     var count: Int,
-    var probability: Int,
+    var probability: Double,
 )
